refactor: 提取 GenUI runtime 基础模块
