+++ src/client/blocks/charts.tsx
@@ -5,7 +5,7 @@
  */
 import { memo, useState } from 'react'
 import css from '../GenuiBlock.module.css'
-import { GENUI_LIMITS } from '../guard.ts'
+import { GENUI_LIMITS } from '../genui-runtime/index.ts'
 import type { GenuiChart, GenuiTable } from '../spec.ts'
 
 export const CHART_COLORS = [
