+++ src/client/genui-runtime/index.ts
@@ -1,0 +1,1 @@
+export { GENUI_LIMITS } from './limits.ts'
